Create lesson3-css-style
In order to create a responsive page , the base of css style file has established with box model.
in this the page has divided into 3 box with diffrent widht percent. 
for the rectangle screen of rgb colors, we have used the absolute positioning style  and as you can see , has included many other properties.