Create lesson4-myScript file
you should download this file and put on the same folder with html and css files and name it myScript. because in the html page we have called this file name. these function have made as DOM to change the css sectors for rectangle screen rgb colors and consist of rgb colors and position and rotation and etc.